+++ Consultoresvs3/Models/ReporteUsuario.cs
@@ -18,7 +18,7 @@ namespace Consultoresvs3.Models
         public DateTime FechaReporte { get; set; }
         [Range(1, 10, ErrorMessage = "{0} las horas deben estar entre {1} y {2} .")]
         [Display(Name = "Horas Trabajadas")]
-        public int HTrabajadas { get; set; }
+        public decimal HTrabajadas { get; set; }
         public int IdServicio { get; set; }
 
         [ForeignKey("IdServicio")]
